+++ aio_taginfo/__init__.py
@@ -30,6 +30,7 @@ __all__ = (
     "key_prevalent_values",
     "key_similar",
     "site_config_geodistribution",
+    "tags_popular",
     "TagInfoError",
 )
 
@@ -38,4 +39,5 @@ from aio_taginfo.api.v4.key.overview import call as key_overview
 from aio_taginfo.api.v4.key.prevalent_values import call as key_prevalent_values
 from aio_taginfo.api.v4.key.similar import call as key_similar
 from aio_taginfo.api.v4.site.config.geodistribution import call as site_config_geodistribution
+from aio_taginfo.api.v4.tags.popular import call as tags_popular
 from aio_taginfo.error import TagInfoError